requirements.txt
future==0.18.2
protobuf==3.17.3
python-whois==0.7.3
six==1.16.0
whois11==0.0.8
